2016-02-03:     1.调整连边权重类型为double

+++ 复杂网络仿真平台-Preussen/Platform/IO/SSTStrategy.cs
@@ -113,7 +113,7 @@ namespace CNSP.Platform.IO
                     intTarget = Convert.ToInt32(strLink[i]) - intNumOffset;
                     if (iNum != intTarget)
                     {
-                        NewNode.AddEdge(intTarget, Convert.ToInt32(strValue[i]));
+                        NewNode.AddEdge(intTarget, Convert.ToDouble(strValue[i]));
                     }
                 }
             }
